+++ devenv/src/devenv/mod.rs
@@ -639,6 +639,8 @@ impl Devenv {
             env,
             bash,
             ignore_process_deps: false,
+            exit_on_idle: false,
+            supervise_processes: true,
         })
     }
 
@@ -1183,6 +1185,8 @@ impl Devenv {
             env: envs,
             bash,
             ignore_process_deps: false,
+            exit_on_idle: false,
+            supervise_processes: true,
         };
 
         let tasks = Tasks::builder(config, verbosity, Arc::clone(&self.shutdown))
@@ -1685,7 +1689,11 @@ impl Devenv {
                 );
                 let result = tasks_runner
                     .process_manager()
-                    .run_foreground(self.shutdown.cancellation_token(), None)
+                    .run_foreground(
+                        self.shutdown.cancellation_token(),
+                        None,
+                        processes::CompletionMode::Persistent,
+                    )
                     .await
                     .map_err(|e| miette!("Process manager error: {}", e));
                 trace!("devenv.up: run_foreground returned");